Move the project from project's repository to 01-awesome-website
To publish a page, Gitub looks for an index in the current folder
and will not find it because the index is placed in a specific project's
folder. For this reason, github does not publish the page, and it turns
out that each project must be a separate repository. That's why
I deleted repository projects and moved the awesome-website to its own.